requirements.txt
google-play-scraper
pandas
numpy
matplotlib
seaborn
wordcloud
transformers
scikit-learn
nltk
psycopg2-binary
python-dotenv